split CryptoTest into HashTest and SignTest
